file transfer / download / upload reorg, part I
