+++ src/cdm_data_loaders/utils/file_transfer/s3/streamer.py
@@ -1,0 +1,3 @@
+"""Stream HTTP resources directly into S3, without buffering to local disk."""
+
+from cdm_data_loaders.utils.s3 import stream_to_s3
